+++ src/server/helpers/json-store.test.ts
@@ -1,22 +1,23 @@
+import { vi, it, describe, expect, Mock } from 'vitest';
 import { open } from './json-store';
 import { asJson, toFile } from './io';
 
-jest.mock('./io', () => ({
-  asJson: jest.fn(),
-  toFile: jest.fn(),
+vi.mock('./io', () => ({
+  asJson: vi.fn(),
+  toFile: vi.fn(),
 }));
 
 describe('JsonStore', () => {
   it('select all without argument gets all items', () => {
     const store = open('foo');
-    (asJson as jest.Mock<any, any>).mockImplementation(() => [1, 2, 3, 4]);
+    (asJson as Mock<any>).mockImplementation(() => [1, 2, 3, 4]);
     const results = store.select();
     expect(results).toEqual([1, 2, 3, 4]);
   });
 
   it('select all with filter gets some items', () => {
     const store = open('foo');
-    (asJson as jest.Mock<any, any>).mockImplementation(() => [1, 2, 3, 4]);
+    (asJson as Mock<any>).mockImplementation(() => [1, 2, 3, 4]);
     const results = store.select((item: any) => item < 3);
     expect(results).toEqual([1, 2]);
   });
@@ -25,7 +26,7 @@ describe('JsonStore', () => {
     const store = open('foo');
     store.switchTo('bar');
     let result = '';
-    (asJson as jest.Mock<any, any>).mockImplementation((name) => {
+    (asJson as Mock<any>).mockImplementation((name: any) => {
       result = name;
       return [];
     });
@@ -36,8 +37,8 @@ describe('JsonStore', () => {
   it('insert changes the target', () => {
     const store = open('foo');
     let stored = undefined;
-    (asJson as jest.Mock<any, any>).mockImplementation(() => [1, 2, 3, 4]);
-    (toFile as jest.Mock<any, any>).mockImplementation((file, result) => {
+    (asJson as Mock<any>).mockImplementation(() => [1, 2, 3, 4]);
+    (toFile as Mock<any>).mockImplementation((file, result: any) => {
       stored = result;
     });
     store.insert(5);
@@ -47,8 +48,8 @@ describe('JsonStore', () => {
   it('delete removes from the target if available', () => {
     const store = open('foo');
     let stored = undefined;
-    (asJson as jest.Mock<any, any>).mockImplementation(() => [1, 2, 3, 4]);
-    (toFile as jest.Mock<any, any>).mockImplementation((file, result) => {
+    (asJson as Mock<any>).mockImplementation(() => [1, 2, 3, 4]);
+    (toFile as Mock<any>).mockImplementation((file, result: any) => {
       stored = result;
     });
     store.delete((item) => item === 4);
@@ -58,8 +59,8 @@ describe('JsonStore', () => {
   it('delete does not remove from the target if not available', () => {
     const store = open('foo');
     let stored = undefined;
-    (asJson as jest.Mock<any, any>).mockImplementation(() => [1, 2, 3, 4]);
-    (toFile as jest.Mock<any, any>).mockImplementation((file, result) => {
+    (asJson as Mock<any>).mockImplementation(() => [1, 2, 3, 4]);
+    (toFile as Mock<any>).mockImplementation((file, result: any) => {
       stored = result;
     });
     store.delete((item) => item === 5);
@@ -69,8 +70,8 @@ describe('JsonStore', () => {
   it('delete does not do anything if source is empty', () => {
     const store = open('foo');
     let stored = false;
-    (asJson as jest.Mock<any, any>).mockImplementation(() => []);
-    (toFile as jest.Mock<any, any>).mockImplementation((file, result) => {
+    (asJson as Mock<any>).mockImplementation(() => []);
+    (toFile as Mock<any>).mockImplementation((file, result) => {
       stored = true;
     });
     store.delete((item) => item === 5);