+++ src/server/management/basics.test.ts
@@ -1,8 +1,10 @@
+import { describe, it, expect } from 'vitest';
 import { resolve } from 'path';
 import { configOf } from './basics';
 
 describe('Management API - Basics', () => {
   it('configOf should return a function', () => {
+    // @ts-ignore
     const handler = configOf(undefined, undefined);
     expect(typeof handler).toBe('function');
   });
@@ -10,8 +12,10 @@ describe('Management API - Basics', () => {
   it('configOf should read the version from the pkg json', () => {
     const version = require(resolve(__dirname, '../../../package.json')).version;
     const config: any = {};
+    // @ts-ignore
     const handler = configOf(undefined, config);
     let result: any;
+    // @ts-ignore
     handler(undefined, {
       json(obj: any) {
         result = obj;
@@ -25,8 +29,10 @@ describe('Management API - Basics', () => {
       name: 'foo',
       directory: 'bar',
     };
+    // @ts-ignore
     const handler = configOf(undefined, config);
     let result: any;
+    // @ts-ignore
     handler(undefined, {
       json(obj: any) {
         result = obj;