+++ CHANGELOG.md
@@ -2,8 +2,11 @@
 
 ## 1.0.0
 
+- Replaced Jest with vitest
 - Updated script injector to support `connected` and `disconnected` events
 - Updated Node.js constraint to 18.17
+- Updated dependencies
+- Added `proxyWebSocket` to exported utilities
 
 ## 0.17.0